Move component out into its own folder
