+++ src/components/src/ProductDisplayItemDetails/productdisplayitem.details.d.ts
