+++ src/components/src/ProductDisplayItemDetails/productdisplayitem.details.tsx
